pyproject.toml
[tool.poetry]
name = "trainaspower"
version = "0.1.0"
description = "Convert TrainAsOne plans to power and upload to Final Surge for use with Stryd pod."
authors = ["Chase Sterling <chase.sterling@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,^3.9"
requests = "^2.24.0"
requests-html = "^0.10.0"
dateparser = "^0.7.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
trainaspower = "trainaspower.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
